% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_indicator.R
\name{geom_indicator}
\alias{geom_indicator}
\title{ggplot2 geom that adds a label to facet panes}
\usage{
geom_indicator(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  parse = FALSE,
  ...,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
ggplot2 geom that adds a label to facet panes
}
